+++ allensdk/__init__.py
@@ -36,7 +36,8 @@
 import logging
 
 
-__version__ = '2.0.2'
+
+__version__ = '2.1.0'
 
 
 try: