+++ allensdk/test/brain_observatory/ecephys/test_ecephys_session_nwb_api.py
@@ -1,5 +1,8 @@
 # most of the tests for this functionality are actually in test_write_nwb
 
+import warnings
+
+import h5py
 import pytest
 import pandas as pd